requirements.txt
asgiref==3.5.2
Django==4.0.5
django-ckeditor==6.4.2
django-cleanup==6.0.0
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-modeltranslation==0.18.2
Pillow==9.1.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
